// enums/employee.enum.ts
export enum Category {
    SC = 'SC',
    ST = 'ST',
    OBC = 'OBC',
    GENERAL = 'GENERAL',
}

export enum EducationQualification {
    UNDER_8 = 'UNDER_8',
    EIGHT = 'EIGHT',
    TEN = 'TEN',
    TWELVE = 'TWELVE',
    GRADUATE = 'GRADUATE',
    POST_GRADUATE = 'POST_GRADUATE',
}

export enum Gender {
    MALE = 'MALE',
    FEMALE = 'FEMALE'
}

export enum EmployeeTitle {
    MR = 'MR',
    MS = 'MS',
}

export enum Status {
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE',
}